Fix issue with untrusted certs
The service worker acts as a programmable cache.  It is not needed for
any functionality.  When the SSL certificate is not trusted,
installation of the service worker will fail.  On some browsers this
breaks JavaScript; in particular the right-cick menu, viewing files and
editing files.

A simple `try`/`catch` block prevents JavaScript from being broken when
the service worker cannot be installed.